chore(rust): refactored kafka service
